// frontend/src/app/swagger/model/instanceCellReadDto.ts
/**
 * S2Retro.Presentation
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InstanceCategoryValueReadDto } from './instanceCategoryValueReadDto';


export interface InstanceCellReadDto { 
    id?: string;
    position?: number;
    textValue?: string | null;
    numberValue?: number | null;
    boolValue?: boolean | null;
    dateValue?: string | null;
    rowId?: string;
    instanceCategoryValueId?: string | null;
    categoryValue?: InstanceCategoryValueReadDto;
}